Update: VSCode generation script syntax
